requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
requests==2.26.0
openai==0.27.0
Flask-WTF==0.15.1
email-validator==1.1.3
python-jose==3.3.0
gunicorn==20.1.0
boto3==1.18.44
